Added query2bib and query2bibtex functions.
query2bib allows user to generate a bibliography from any query that includes a ref column. Uses the prewritten bibliographies on ADS and writes them all to an array. Bibcodes that do not have an ads bibliography will be presented after the function runs so users can inspect the codes or look for them manually.

query2bibtex is query2bib, but instead of returning the bibliography and an array, generates a bibtex file for the user.